create togglable list/card views for roster view
