+++ client/src/views/Dashboard/Dashboard.less
@@ -38,6 +38,9 @@
         margin: auto;
         padding-left: 0;
         padding-right: 0;
+        &:hover {
+            width: 11vw;
+        }
     }
 }
