+++ client/src/views/Roster/Roster.less
@@ -9,10 +9,12 @@
   text-align: left;
 }
 
+
+// Table View
 #table-container {
   width: 80%;
   padding-top: 5vh;
-  margin: auto;
+  margin: 2vh auto 0 auto;
 }
 
 .ant-table {
@@ -23,4 +25,47 @@
 
 .ant-table-row {
   background-color: #colors[tertiary];
+}
+
+
+// Card View
+#card-container {
+  width: 80vw;
+  padding: 0 0 5vh 0;
+  margin: 0 auto 0 auto;
+  display: inline-grid;
+  grid-template-columns: auto auto;
+}
+
+#card {
+  text-align: left;
+  width: 35vw;
+  height: 45vh;
+  margin: 2vh auto 0 auto;
+  background: #colors[tertiary];
+  color: #colors[text-primary];
+  border: solid 4px;
+  border-color: #colors[secondary];
+  border-radius: 10px;
+}
+
+#card-button-container {
+  margin: 5% auto 0 auto;
+  button {
+    width: 10vw;
+    font-size: 1em;
+    border-radius: 4px;
+    margin: auto;
+    padding-left: 0;
+    padding-right: 0;
+    &:hover {
+      width: 11vw;
+    }
+  }
+}
+
+#card-content-container {
+  position: relative;
+  bottom: 0;
+  margin-bottom: 1vh;
 }
